+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/action/MultipleSubtasksConfigurationAction.java
@@ -3,7 +3,11 @@ package de.codescape.jira.plugins.multiplesubtasks.action;
 import com.atlassian.jira.security.request.RequestMethod;
 import com.atlassian.jira.security.request.SupportedMethods;
 import com.atlassian.jira.web.action.JiraWebActionSupport;
+import de.codescape.jira.plugins.multiplesubtasks.model.TemplateSortOrder;
 import de.codescape.jira.plugins.multiplesubtasks.service.MultipleSubtasksConfigurationService;
+
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * This action allows to configure Multiple Subtasks for Jira.
@@ -28,6 +32,20 @@ public class MultipleSubtasksConfigurationAction extends JiraWebActionSupport {
      */
     public String getTemplatesPerProjectKey() {
         return MultipleSubtasksConfigurationService.TEMPLATES_PER_PROJECT;
+    }
+
+    /**
+     * Return the key for the sort order of templates.
+     */
+    public String getTemplatesSortOrderKey() {
+        return MultipleSubtasksConfigurationService.TEMPLATES_SORT_ORDER;
+    }
+
+    /**
+     * Return the options for the sort order of templates.
+     */
+    public List<TemplateSortOrder> getTemplatesSortOrderOptions() {
+        return Arrays.asList(TemplateSortOrder.values());
     }
 
 }